Changed temperature an humidity type to double

+++ Backend/SaunaMonitor/Models/measurement.cs
@@ -3,9 +3,10 @@ namespace SaunaMonitor.Models
     public class Measurement
     {
         public int Id { get; set; }
-        public float TemperatureC { get; set; }
-        public float Humidity { get; set; }
+        public double TemperatureC { get; set; }
+        public double Humidity { get; set; }
         public DateTime ReadingDate { get; set; }
+        
         
     }
